Sloppy Fix + Numbers Font loaded on init
The sloppy fix is for when the user starts on a film card set and moves to addition/subtraction.
What happens in the background is that the defaults are checked for add/sub. If they arent found a hard refresh is called on the scene.
However, as the game deals with the rest of the function if will spawn a failure and several warnings.
It successfully restarts the scene, but as said, its so sloppy.
Example is in the code to see what I mean.